requirements.txt
pillow
pygetwindow
pyautogui 